endpoint: Move event start/stop to events.go
This is a pure code move, no functional changes.
