--- pkg/endpoint/manager.go
@@ -1,64 +1,0 @@
-// Copyright 2016-2019 Authors of Cilium
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package endpoint
-
-import (
-	"fmt"
-
-	"github.com/cilium/cilium/pkg/eventqueue"
-	"github.com/cilium/cilium/pkg/logging/logfields"
-	"github.com/cilium/cilium/pkg/option"
-)
-
-// Start assigns a Cilium Endpoint ID to the endpoint and prepares it to
-// receive events from other subsystems.
-//
-// The endpoint must not already be exposed via the endpointmanager prior to
-// calling Start(), as it assumes unconditional access over the Endpoint
-// object.
-func (e *Endpoint) Start(id uint16) {
-	// No need to check liveness as an endpoint can only be deleted via the
-	// API after it has been inserted into the manager.
-	// 'e.ID' written below, read lock is not enough.
-	e.unconditionalLock()
-	defer e.unlock()
-
-	e.ID = id
-	e.UpdateLogger(map[string]interface{}{
-		logfields.EndpointID: e.ID,
-	})
-
-	// Start goroutines that are responsible for handling events.
-	e.startRegenerationFailureHandler()
-	if e.eventQueue == nil {
-		e.InitEventQueue()
-	}
-	e.eventQueue.Run()
-	e.getLogger().Info("New endpoint")
-}
-
-// InitEventQueue initializes the endpoint's event queue. Note that this
-// function does not begin processing events off the queue, as that's left up
-// to the caller to call Expose in order to allow other subsystems to access
-// the endpoint. This function assumes that the endpoint ID has already been
-// allocated!
-//
-// Having this be a separate function allows us to prepare
-// the event queue while the endpoint is being validated (during restoration)
-// so that when its metadata is resolved, events can be enqueued (such as
-// visibility policy and bandwidth policy).
-func (e *Endpoint) InitEventQueue() {
-	e.eventQueue = eventqueue.NewEventQueueBuffered(fmt.Sprintf("endpoint-%d", e.ID), option.Config.EndpointQueueSize)
-}
